requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
networkx
scipy
google-cloud-bigquery
google-auth
pydantic
